refactor scrollable info sections into separate components
