+++ app/(index)/Scene.tsx
@@ -1,13 +1,13 @@
 import * as THREE from "three";
 
-import { motion } from "framer-motion";
 import { useRef } from "react";
 
-import { Scroll, useScroll, Svg, Text, Html, Center } from "@react-three/drei";
+import { Scroll, useScroll, Svg, Center } from "@react-three/drei";
 import { useFrame } from "@react-three/fiber";
 
-import { externalRoutes } from "../routes";
-
+import DescriptionSection from "./DescriptionSection";
+import DirectorSection from "./DirectorSection";
+import WatchNowSection from "./WatchNowSection";
 import Xbot from "./Xbot";
 
 interface Props {
@@ -17,11 +17,6 @@ const Scene = (props: Props) => {
   const { onCalculateXbotHeight } = props;
 
   const xbotRef = useRef<THREE.Group | null>(null);
-
-  const directorTextRef = useRef<THREE.Group | null>(null);
-  const descTextRef = useRef<THREE.Group | null>(null);
-  const watchTextRef = useRef<THREE.Group | null>(null);
-
   const scroll = useScroll();
 
   useFrame(() => {
@@ -30,21 +25,6 @@ const Scene = (props: Props) => {
         THREE.MathUtils.degToRad(-90) -
         THREE.MathUtils.degToRad(scroll.offset * 360);
       xbotRef.current.rotation.y = rotationOffset;
-    }
-
-    if (directorTextRef.current) {
-      const translationOffset = scroll.range(15 / 100, 74 / 100);
-      directorTextRef.current.position.y = 7.2 + 2 * translationOffset;
-    }
-
-    if (descTextRef.current) {
-      const translationOffset = scroll.range(38 / 100, 56 / 100);
-      descTextRef.current.position.y = 6.8 - 3.2 * translationOffset;
-    }
-
-    if (watchTextRef.current) {
-      const translationOffset = scroll.range(60 / 100, 40 / 100);
-      watchTextRef.current.position.y = 4.2 - 2.6 * translationOffset;
     }
   });
 
@@ -64,122 +44,9 @@ const Scene = (props: Props) => {
         />
       </Center>
 
-      <group ref={directorTextRef} position={[-4.5, 7.2, -3]}>
-        <Text
-          position={[0, 0.62, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={0.18}
-          letterSpacing={0.55}
-        >
-          Created by
-        </Text>
-        <Text
-          position={[0, 0, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={0.6}
-          letterSpacing={0.2}
-        >
-          Tim
-        </Text>
-        <Text
-          position={[0, -0.72, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={1}
-        >
-          Miller
-        </Text>
-      </group>
-
-      <group ref={descTextRef} position={[0.4, 6.8, 2]}>
-        <Text
-          position={[0, 0.27, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={0.23}
-          letterSpacing={0.2}
-        >
-          Sci-Fi
-        </Text>
-        <Text
-          position={[0, 0, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={0.37}
-        >
-          Television
-        </Text>
-        <Text
-          position={[0, -0.35, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={0.07}
-          lineHeight={1.2}
-          letterSpacing={0.05}
-        >
-          Terrifying creatures, wicked surprised and {`\n`}
-          dark comedy converge in this NSFW anthology {`\n`}
-          of animated stories.
-        </Text>
-      </group>
-
-      <group ref={watchTextRef} position={[-1.7, 0, 2]}>
-        <Text
-          position={[0, 0.23, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={0.07}
-          letterSpacing={0.55}
-        >
-          Watch online
-        </Text>
-        <Text
-          position={[0, 0, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={0.23}
-          letterSpacing={0.2}
-        >
-          Only on
-        </Text>
-        <Text
-          onClick={() => console.log("clicked")}
-          position={[0, -0.27, 0]}
-          anchorX="left"
-          color="white"
-          font="/fonts/Bebas-Regular.ttf"
-          fontSize={0.37}
-        >
-          Netflix
-        </Text>
-        <Html
-          transform
-          zIndexRange={[9, 0]}
-          scale={0.15}
-          position={[0.29, -0.65, 0]}
-        >
-          <motion.a
-            href={externalRoutes.netflixLDRTitle}
-            whileHover={{
-              backgroundColor: "var(--color-botred)",
-              color: "#fff",
-            }}
-            className="border-botgray text-botgray rounded-full border-2 border-solid bg-transparent px-10 py-3"
-          >
-            Watch now
-          </motion.a>
-        </Html>
-      </group>
+      <DirectorSection />
+      <DescriptionSection />
+      <WatchNowSection />
     </Scroll>
   );
 };
